# tunnel/openapi.yaml
openapi: 3.1.0
info:
  title: Prime Tunnel API
  version: 0.2.0
  description: Secure tunnel API to read repo status/logs/files and (optionally) run allowlisted commands.
servers:
  - url: http://127.0.0.1:8787
    description: Local Prime Tunnel server
  - url: https://{tunnelHost}
    description: Cloudflare tunnel (set to the latest printed host)
    variables:
      tunnelHost:
        default: CHANGE_ME.trycloudflare.com

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthResponse:
      type: object
      required: [ok, time, repoRoot, allowExec]
      properties:
        ok: { type: boolean }
        time: { type: string }
        repoRoot: { type: string }
        allowExec: { type: boolean }

    GitStatusResponse:
      type: object
      required: [ok, code, signal, stdout, stderr, stdout_truncated, stderr_truncated]
      properties:
        ok: { type: boolean }
        code: { type: integer }
        signal: { type: ['string', 'null'] }
        stdout: { type: string }
        stderr: { type: string }
        stdout_truncated: { type: boolean }
        stderr_truncated: { type: boolean }

    LogsTailResponse:
      type: object
      required: [ok, path, lines]
      properties:
        ok: { type: boolean }
        path: { type: string }
        lines:
          type: array
          items: { type: string }

    FileReadRequest:
      type: object
      required: [path]
      properties:
        path: { type: string }
        startLine: { type: integer, minimum: 1 }
        endLine: { type: integer, minimum: 1 }

    FileReadResponse:
      type: object
      required: [ok, path, startLine, endLine, content, totalLines]
      properties:
        ok: { type: boolean }
        path: { type: string }
        startLine: { type: integer }
        endLine: { type: integer }
        content: { type: string }
        totalLines: { type: integer }

    ExecRequest:
      type: object
      required: [cmd]
      properties:
        cmd: { type: string }
        args:
          type: array
          items: { type: string }
        cwdRel: { type: string }

    ExecResponse:
      type: object
      required:
        [ok, code, signal, stdout, stderr, stdout_truncated, stderr_truncated, cmd, args, cwdRel]
      properties:
        ok: { type: boolean }
        code: { type: integer }
        signal: { type: ['string', 'null'] }
        stdout: { type: string }
        stderr: { type: string }
        stdout_truncated: { type: boolean }
        stderr_truncated: { type: boolean }
        cmd: { type: string }
        args:
          type: array
          items: { type: string }
        cwdRel: { type: string }

paths:
  /health:
    get:
      operationId: primeTunnelHealth
      summary: Health check
      security: [] # no auth
      responses:
        '200':
          description: Health JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /git/status:
    get:
      operationId: primeTunnelGitStatus
      summary: Get git status porcelain
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Git status result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatusResponse'

  /logs/tail:
    get:
      operationId: primeTunnelLogsTail
      summary: Tail a log file (default .tunnel/terminal_prime.log)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: path
          required: false
          schema: { type: string }
        - in: query
          name: n
          required: false
          schema: { type: integer, minimum: 1, maximum: 5000 }
      responses:
        '200':
          description: Tail results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsTailResponse'

  /file/read:
    post:
      operationId: primeTunnelFileRead
      summary: Read file line range within repo root
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileReadRequest'
      responses:
        '200':
          description: File content with metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileReadResponse'

  /exec:
    post:
      operationId: primeTunnelExec
      summary: Run allowlisted command (disabled unless ALLOW_EXEC=1)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecRequest'
      responses:
        '200':
          description: Exec result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecResponse'